Remove bad softlayer-cpi release reference
Multiple versions of 31 were added - one under the name
bosh-softlayer-ng-cpi and the other as bosh-softlayer-cpi. In theory
BOSH supports this, but bosh.io uses dynamic glob matching to find
the tarball file on the release name (e.g. `*-31/source.meta4`), so
this fails.

Since this version was removed upstream *and* this breaks bosh.io on
their release page *and* it would probably take too long to fix
`bosh-io/web` to behave correctly in this case... manually removing
these references. Normally, release authors should simply create new
versions after a broken release.

The underlying error around this was masked by a different bug in
bosh-io/web which was fixed in bosh-io/web@845b999b81.

https://github.com/bosh-io/web/issues/38